Optionnal responsive.less
This file is only available for bootstrap 2.x

+++ src/Sanpi/TwitterBootstrap/Composer/ScriptHandler.php
@@ -40,8 +40,10 @@ class ScriptHandler
         $css = $lessc->compileFile("$bootstrapDir/less/bootstrap.less");
         file_put_contents("$webDir/css/bootstrap.css", $css);
 
-        $css = $lessc->compileFile("$bootstrapDir/less/responsive.less");
-        file_put_contents("$webDir/css/bootstrap-responsive.css", $css);
+        if (is_file("$bootstrapDir/less/responsive.less")) {
+            $css = $lessc->compileFile("$bootstrapDir/less/responsive.less");
+            file_put_contents("$webDir/css/bootstrap-responsive.css", $css);
+        }
 
         foreach (glob("$bootstrapDir/js/*.js") as $src) {
             $dst = "$webDir/js/" . basename($src);
